refactor(packet): split ethernet helpers
